backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart>=0.0.7
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
google-genai>=1.0.0
openai>=1.3.5
numpy>=1.25.0
pandas>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pydantic>=2.5.0
aiohttp>=3.9.1
scikit-learn>=1.3.0
nltk>=3.8.1